Move semantic validation from ReadBuilder to Read
This commit moves most of the semantic validation previously done in `dimacs.ReadBuilder` into `cnfBuilder`. This makes `dimacs.ReadBuilder` simpler as it now focuses solely on verifying that each line is semantically correct rather than verifying that the whole file is. For example, `dimacs.ReadBuilder` doesn't take care of verifying that the file contains exactly one problem line anymore. Rather, it is the responsibility of the Builder to decide whether this is OK or not.