Treat '%' as an end of file marker
Char '%' indicates the end of the clause list. Though, some file might still contain semantically valid clause lines after '%'. For example, instances of the "uf20-91" family have an empty clause after '%'.

Treating '%' as the end of file (rather than ignoring it) prevents incorrectly adding empty clauses to the builder. This is consistent with the behavior of other DIMACS parsers such as the one in MiniSat.